refactor(define-common): extract updateVisibilityIfSublevel()
